requirements.txt
flask==2.0.1
python-docx==0.8.11
PyPDF2==3.0.1
pdfplumber==0.7.4
python-magic==0.4.27
tqdm==4.65.0
python-dotenv==0.19.0
SQLAlchemy==1.4.23 